+++ R/job.R
@@ -35,7 +35,7 @@ NULL
 #' job <- eplus_job(idf, epw)
 #' job$version()
 #' job$path(type = c("all", "idf", "epw"))
-#' job$run(wait = TRUE, force = FALSE)
+#' job$run(wait = TRUE, force = FALSE, echo = wait)
 #' job$kill()
 #' job$status()
 #' job$errors(info = FALSE)
@@ -43,6 +43,8 @@ NULL
 #' job$locate_output(suffix = ".err", strict = TRUE)
 #' job$list_table()
 #' job$read_table(name)
+#' job$read_rdd()
+#' job$read_mdd()
 #' job$report_data_dict()
 #' job$report_data(key_value = NULL, name = NULL, year = NULL, tz = "UTC", case = "auto",
 #'                 all = FALSE, period = NULL, month = NULL, day = NULL, hour = NULL,
@@ -70,11 +72,11 @@ NULL
 #' * `epw`: Path to an local EnergyPlus EPW file or an [Epw] object.
 #' * `type`: If `"all"`, both the [Idf] path and [Epw] path are returned. If
 #'   `"idf"`, only IDF path is returned. If `"epw"`, only EPW path is returned.
-#'   Default: `"all"`.
+#'   Default: `"all"`. If `epw` is `NULL`, `NA` is returned for EPW path.
 #'
 #' @section Run:
 #' ```
-#' job$run(wait = TRUE, force = FALSE)
+#' job$run(wait = TRUE, force = FALSE, echo = wait)
 #' job$kill()
 #' job$status()
 #' job$errors(info = FALSE)
@@ -110,6 +112,8 @@ NULL
 #'   EnergyPlus standard output (stdout) and error (stderr) is printed to
 #'   R console. If `FALSE`, simulation will be run in a background process.
 #'   Default: `TRUE`.
+#' * `echo`: Only applicable when `wait` is `TRUE`. Whether to show standard
+#'   output and error from EnergyPlus. Default: same as `wait`.
 #' * `force`: Only applicable when the last job runs with `wait` equals
 #'   to `FALSE` and is still running. If `TRUE`, current running job is
 #'   forced to stop and a new one will start. Default: `FALSE`.
@@ -120,7 +124,9 @@ NULL
 #' job$output_dir(open = FALSE)
 #' job$locate_output(suffix = ".err", strict = TRUE)
 #' job$list_table()
-#' job$read_table(table)
+#' job$read_table(name)
+#' job$read_rdd()
+#' job$read_mdd()
 #' job$report_data_dict()
 #' job$report_data(key_value = NULL, name = NULL, year = NULL, tz = "UTC", case = "auto",
 #'                 all = FALSE, period = NULL, month = NULL, day = NULL, hour = NULL,
@@ -137,8 +143,12 @@ NULL
 #'
 #' `$list_table()` returns all available table and view names in the SQLite file.
 #'
-#' `$read_table()` takes a valid `table` name of those from `$list_table()` and
+#' `$read_table()` takes a valid table name of those from `$list_table()` and
 #' returns that table data in a [data.table][data.table::data.table()] format.
+#'
+#' `$read_rdd()` and `$read_mdd()` return the core data of Report Data
+#' Dictionary (RDD) file and Meter Data Dictionary (MDD) file respectively. For
+#' details, please see [read_rdd()].
 #'
 #' `$report_data_dict()` returns a [data.table][data.table::data.table()] which
 #' contains all information about report data. For details on the meaning of
@@ -388,7 +398,10 @@ NULL
 #' please see [EplusJob].
 #'
 #' @param idf A path to an local EnergyPlus IDF file or an `Idf` object.
-#' @param epw A path to an local EnergyPlus EPW file or an `Epw` object.
+#' @param epw A path to an local EnergyPlus EPW file or an `Epw` object. `epw`
+#' can also be `NULL` which will force design-day-only simulation when
+#' [`$run()`][EplusJob] method is called. Note this needs at least one
+#' `Sizing:DesignDay` object exists in the [Idf].
 #' @return An `EplusJob` object.
 #' @examples
 #' if (is_avail_eplus(8.8)) {
@@ -418,7 +431,7 @@ EplusJob <- R6::R6Class(classname = "EplusJob", cloneable = FALSE,
     public = list(
 
         # INITIALIZE {{{
-        initialize = function (idf, epw, eplus_ver = NULL) {
+        initialize = function (idf, epw = NULL, eplus_ver = NULL) {
 
             if (is_idf(idf)) {
                 private$m_path_idf <- idf$path()
@@ -449,34 +462,38 @@ EplusJob <- R6::R6Class(classname = "EplusJob", cloneable = FALSE,
 
             private$m_path_idf <- normalizePath(private$m_path_idf, mustWork = TRUE)
 
-            if (is_epw(epw)) {
-                private$m_path_epw <- epw$path()
-                if (is.null(private$m_path_epw)) {
-                    abort("error_epw_not_local",
-                        paste0(
-                            "The Epw object is not created from local file. ",
-                            "Please give save it to disk before run."
-                        )
-                    )
-                }
-
-                if (epw$is_unsaved()) {
-                    abort("error_epw_not_saved",
-                        paste0("Epw has been modified since read or last saved. ",
-                            "Please save Epw using $save() before run."
-                        )
-                    )
-                }
+            if (is.null(epw)) {
+                private$m_path_epw <- NULL
             } else {
-                assert(is_string(epw))
-                private$m_path_epw <- epw
-            }
+                if (is_epw(epw)) {
+                    private$m_path_epw <- epw$path()
+                    if (is.null(private$m_path_epw)) {
+                        abort("error_epw_not_local",
+                            paste0(
+                                "The Epw object is not created from local file. ",
+                                "Please give save it to disk before run."
+                            )
+                        )
+                    }
 
-            if (!file.exists(private$m_path_epw)) {
-                abort("error_epw_not_exist", "Input epw file does not exists.")
-            }
+                    if (epw$is_unsaved()) {
+                        abort("error_epw_not_saved",
+                            paste0("Epw has been modified since read or last saved. ",
+                                "Please save Epw using $save() before run."
+                            )
+                        )
+                    }
+                } else {
+                    assert(is_string(epw))
+                    private$m_path_epw <- epw
+                }
 
-            private$m_path_epw <- normalizePath(private$m_path_epw, mustWork = TRUE)
+                if (!file.exists(private$m_path_epw)) {
+                    abort("error_epw_not_exist", "Input epw file does not exists.")
+                }
+
+                private$m_path_epw <- normalizePath(private$m_path_epw, mustWork = TRUE)
+            }
 
             # get Idf version
             if (!is.null(eplus_ver)) {
@@ -514,8 +531,8 @@ EplusJob <- R6::R6Class(classname = "EplusJob", cloneable = FALSE,
         path = function (type = c("all", "idf", "epw"))
             job_path(self, private, type),
 
-        run = function (wait = TRUE, force = FALSE)
-            job_run(self, private, wait = wait, force),
+        run = function (wait = TRUE, force = FALSE, echo = wait)
+            job_run(self, private, wait = wait, force, echo),
 
         kill = function ()
             job_kill(self, private),
@@ -534,6 +551,12 @@ EplusJob <- R6::R6Class(classname = "EplusJob", cloneable = FALSE,
 
         list_table = function ()
             job_list_table(self, private),
+
+        read_rdd = function ()
+            job_read_rdd(self, private),
+
+        read_mdd = function ()
+            job_read_mdd(self, private),
 
         read_table = function (name)
             job_read_table(self, private, name),
@@ -586,14 +609,15 @@ job_version <- function (self, private) {
 job_path <- function (self, private, type = c("all", "idf", "epw")) {
     type <- match.arg(type)
 
+    path_epw <- private$m_path_epw %||% NA_character_
     switch(type,
-        all = c(idf = private$m_path_idf, epw = private$m_path_epw),
-        idf = private$m_path_idf, epw = private$m_path_epw
+        all = c(idf = private$m_path_idf, epw = path_epw),
+        idf = private$m_path_idf, epw = path_epw
     )
 }
 # }}}
 # job_run {{{
-job_run <- function (self, private, wait = TRUE, force = FALSE) {
+job_run <- function (self, private, wait = TRUE, force = FALSE, echo = wait) {
     # check if the model is still running
     old <- private$m_job
     if (!is.null(old)) {
@@ -617,7 +641,9 @@ job_run <- function (self, private, wait = TRUE, force = FALSE) {
     private$m_log$killed <- NULL
 
     private$m_job <- run_idf(private$m_path_idf, private$m_path_epw,
-        output_dir = NULL, echo = wait, wait = wait, eplus = private$m_version)
+        output_dir = NULL, echo = echo, wait = wait, eplus = private$m_version,
+        design_day = is.null(private$m_path_epw)
+    )
 
     if (wait) private$m_log$end_time <- Sys.time()
     self
@@ -766,7 +792,11 @@ job_locate_output <- function (self, private, suffix = ".err", strict = TRUE, mu
 
     }
 
-    if (must_exist) assert(file.exists(out))
+    if (must_exist) {
+        assert(file.exists(out), msg = paste0("File ", surround(out), " does not exists."),
+            err_type = "error_file_not_exist"
+        )
+    }
 
     out
 }
@@ -784,18 +814,50 @@ job_output_errors <- function (self, private, info = FALSE) {
 job_sql_path <- function (self, private) {
     path_sql <- job_locate_output(self, private, ".sql", must_exist = FALSE)
     if (!file.exists(path_sql))
-        stop("Simulation SQL output does not exists. ",
+        abort("error_sql_not_exist", paste0(
+             "Simulation SQL output does not exist. ",
              "eplusr uses the EnergyPlus SQL output for extracting simulation outputs. ",
              "Please add an object in `Output:SQLite` with `Option Type` value of `SimpleAndTabular` ",
              "and run the Idf again. It is recommended to first read that IDF file using `read_idf()` ",
              "and then use `$run()` method in Idf class by doing `idf$run()` ",
-             "which automatically handle this.", call. = FALSE)
+             "which automatically handle this."
+        ))
     path_sql
+}
+# }}}
+# job_rdd_path {{{
+job_rdd_path <- function (self, private, type = c("rdd", "mdd")) {
+    type <- match.arg(type)
+    path <- job_locate_output(self, private, paste0(".", type), must_exist = FALSE)
+    name <- switch(type,
+        rdd = "Report Data Dictionary (RDD) file",
+        mdd = "Meter Data Dictionary (MDD) file"
+    )
+    if (!file.exists(path))
+        assert("error_rdd_not_exist", paste0(
+             name, " does not exist. ",
+             "Please add an object in `Output:VariableDictionary` class ",
+             "and run the Idf again. It is recommended to first read that IDF file using `read_idf()` ",
+             "and then use `$run()` method in Idf class by doing `idf$run()` ",
+             "which automatically handle this."
+        ))
+
+    path
 }
 # }}}
 # job_list_table {{{
 job_list_table <- function (self, private) {
     list_sql_table(job_sql_path(self, private))
+}
+# }}}
+# job_read_rdd {{{
+job_read_rdd <- function (self, private) {
+    read_rdd(job_rdd_path(self, private, "rdd"))
+}
+# }}}
+# job_read_mdd {{{
+job_read_mdd <- function (self, private) {
+    read_mdd(job_rdd_path(self, private, "mdd"))
 }
 # }}}
 # job_read_table {{{