+++ tests/testthat/test_param.R
@@ -86,15 +86,22 @@ test_that("Parametric methods", {
     # Can detect if models are modified before running
     model2 <- param$models()$set_infil_rate_2
     model2$Output_Variable <- NULL
-    expect_warning(param$run(), class = "warn_param_modified")
+    expect_warning(param$run(echo = FALSE), class = "warn_param_modified")
 
     dir_nms <- paste0("set_infil_rate_", 1:5)
     param$apply_measure(set_infil_rate, seq(0, 4, by = 1), .names = NULL)
     # can run the simulation and get status of simulation
-    expect_equal(
-        {param$run(dir = NULL); param$status()},
+    expect_equal({param$run(dir = NULL, echo = FALSE); status <- param$status(); names(status)},
+        c("run_before", "alive", "terminated", "successful", "changed_after", "job_status")
+    )
+    expect_equal(status[c("run_before", "alive", "terminated", "successful", "changed_after")],
         list(run_before = TRUE, alive = FALSE, terminated = FALSE,
             successful = TRUE, changed_after = FALSE
+        )
+    )
+    expect_equal(names(status$job_status),
+        c("index", "status", "idf", "epw", "exit_status", "start_time", "end_time",
+          "energyplus", "output_dir", "stdout", "stderr"
         )
     )
     # }}}